Refacto/vqa canvas (#473)
* build: remove unused dependencies of vqa canvas

* refacto: create features directory and add annotateItem & addQuestion features

* refacto: use p element instead of div for basic text component

* fix: post rebase

* fix: add empty inference_metadata when creating message
