+++ ui/components/vqaCanvas/src/features/addQuestion/components/index.ts
@@ -1,0 +1,1 @@
+export { default as AddQuestionButton } from "./AddQuestionButton.svelte";
