+++ ui/components/vqaCanvas/src/features/addQuestion/types/index.ts
@@ -1,0 +1,7 @@
+/*-------------------------------------
+Copyright: CEA-LIST/DIASI/SIALV/LVA
+Author : pixano@cea.fr
+License: CECILL-C
+-------------------------------------*/
+
+export * from "./StoreQuestionEvent";
