+++ ui/components/vqaCanvas/src/features/annotateItem/assets/icons/completed-question.png
