+++ ui/components/vqaCanvas/src/features/annotateItem/assets/icons/pending-question.png
