+++ ui/components/vqaCanvas/src/features/annotateItem/components/QuestionHeader.svelte
