+++ ui/components/vqaCanvas/src/features/annotateItem/components/index.ts
