+++ ui/components/vqaCanvas/src/features/annotateItem/types/ContentChangeEvent.ts
