+++ ui/components/vqaCanvas/src/features/annotateItem/utils/answerChoicesToCheckboxsState.ts
