+++ ui/components/vqaCanvas/src/features/annotateItem/utils/checkboxsStateToAnswerChoices.ts
