+++ ui/components/vqaCanvas/src/features/annotateItem/utils/deserializeMessageContent.ts
