+++ ui/components/vqaCanvas/src/features/annotateItem/utils/isQuestionCompleted.ts
