+++ ui/components/vqaCanvas/src/features/annotateItem/utils/serializeMessageContent.ts
